% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{callImaiIri}
\alias{callImaiIri}
\title{Find a minimal piecewise linear function that ensures that all integer points below it lie on or below the input-defined one.
Return the breakpoints of this function, which are guaranteed to be as few as possible in number (see the original paper).
The width specifies the width of the "tunnel" used; for best results, it should be close to 1, but strictly smaller than it.
Use the standard Imai-Iri algorithm of given width.}
\usage{
callImaiIri(extremePoints, width = WIDTH, CPP = TRUE, fname = "Test.csv")
}
\description{
Find a minimal piecewise linear function that ensures that all integer points below it lie on or below the input-defined one.
Return the breakpoints of this function, which are guaranteed to be as few as possible in number (see the original paper).
The width specifies the width of the "tunnel" used; for best results, it should be close to 1, but strictly smaller than it.
Use the standard Imai-Iri algorithm of given width.
}
