% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaiIri.R
\name{checkAcute}
\alias{checkAcute}
\title{Auxiliary function for the Imai-Iri algorithm}
\usage{
checkAcute(point1, point2, point3, sign, tol = TOL)
}
\description{
Auxiliary function for the Imai-Iri algorithm
}
