% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{computeAgreements}
\alias{computeAgreements}
\title{Compute the matrix of pairwise agreements between columns of matrix1 and matrix2, assumed to have the same number of rows}
\usage{
computeAgreements(matrix1, matrix2)
}
\description{
Compute the matrix of pairwise agreements between columns of matrix1 and matrix2, assumed to have the same number of rows
}
