% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{computeBestPValues}
\alias{computeBestPValues}
\title{Compute the column with best metric in matrix2 for each column of matrix1, returning both its value (arg1) and name (arg2).
The matrices are assumed (without checking) to have the same number of rows and to have binary values (0/1).
If fisher = TRUE use the Fisher exact p-value, otherwise chi-squared log p-value.}
\usage{
computeBestPValues(matrix1, matrix2, fisher = TRUE)
}
\description{
Compute the column with best metric in matrix2 for each column of matrix1, returning both its value (arg1) and name (arg2).
The matrices are assumed (without checking) to have the same number of rows and to have binary values (0/1).
If fisher = TRUE use the Fisher exact p-value, otherwise chi-squared log p-value.
}
