% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{computeCovariances}
\alias{computeCovariances}
\title{Compute the matrix of pairwise covariances between columns of matrix1 and matrix2, assumed to have the same number of rows
If scaleUp = TRUE, do not normalize by the number of non-missing pairs.}
\usage{
computeCovariances(matrix1, matrix2, scaleUp = TRUE)
}
\description{
Compute the matrix of pairwise covariances between columns of matrix1 and matrix2, assumed to have the same number of rows
If scaleUp = TRUE, do not normalize by the number of non-missing pairs.
}
