% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{deleteRepeatedRows}
\alias{deleteRepeatedRows}
\title{This function deletes any repeated rows of the input matrix}
\usage{
deleteRepeatedRows(inputMatrix)
}
\description{
This function deletes any repeated rows of the input matrix
}
