% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{deleteZeroColumns}
\alias{deleteZeroColumns}
\title{Delete any all-zero columns of a logical input matrix}
\usage{
deleteZeroColumns(inputMatrix)
}
\description{
Delete any all-zero columns of a logical input matrix
}
