% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\docType{data}
\name{DELTA}
\alias{DELTA}
\title{ILP and CPLEX solver settings in ILP.R
minimal amount by which to exceed the trivial bound}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DELTA
}
\description{
ILP and CPLEX solver settings in ILP.R
minimal amount by which to exceed the trivial bound
}
\keyword{datasets}
