% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{extractMinimalRows}
\alias{extractMinimalRows}
\title{Extract the minimal rows of a logical input matrix with respect to set inclusion}
\usage{
extractMinimalRows(inputMatrix)
}
\description{
Extract the minimal rows of a logical input matrix with respect to set inclusion
}
