% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaiIri.R
\name{findIntersection}
\alias{findIntersection}
\title{Find the intersection between two lines, each specified by a pair of points (2 x 2 matrices; row = point, columns = coords)
Note: does not check that the intersection point lies between the starting and ending points of each segment!}
\usage{
findIntersection(endPoints1, endPoints2)
}
\description{
Find the intersection between two lines, each specified by a pair of points (2 x 2 matrices; row = point, columns = coords)
Note: does not check that the intersection point lies between the starting and ending points of each segment!
}
