% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaiIri.R
\name{findSlopeAndIntercept}
\alias{findSlopeAndIntercept}
\title{Find the slope and y-intercept of a line given by a 2 x 2 matrix, whose rows are points and columns are x and y coordinates
Note: special cases are defined in the dependent functions for horizontal lines (slope = 0) and vertical lines (slope = Inf)}
\usage{
findSlopeAndIntercept(endPoints)
}
\description{
Find the slope and y-intercept of a line given by a 2 x 2 matrix, whose rows are points and columns are x and y coordinates
Note: special cases are defined in the dependent functions for horizontal lines (slope = 0) and vertical lines (slope = Inf)
}
