% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{findSubsetRows}
\alias{findSubsetRows}
\title{Find any rows of the input matrix which are subsets of the input vector; both are assumed to be logical}
\usage{
findSubsetRows(inputMatrix, inputVector)
}
\description{
Find any rows of the input matrix which are subsets of the input vector; both are assumed to be logical
}
