% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaiIri.R
\name{getSignedAngle}
\alias{getSignedAngle}
\title{Auxiliary function for the Imai-Iri algorithm}
\usage{
getSignedAngle(point1, point2, point3, sign)
}
\description{
Auxiliary function for the Imai-Iri algorithm
}
