% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\name{mainDriver}
\alias{mainDriver}
\title{Main driver}
\usage{
mainDriver(
  inputFile,
  type = "CNF",
  objective = "agreement",
  K = 3,
  L = 3,
  complement = FALSE,
  ext = stringr::str_sub(inputFile, -4),
  index = 1L,
  extremeValue = log(MAX_P),
  outputSummary = TRUE,
  outputPhenotype = FALSE,
  outputAssociations = FALSE,
  saveOutput = FALSE
)
}
\description{
Main driver
}
