% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILP.R
\name{makeControlLines}
\alias{makeControlLines}
\title{This function produces instructions controlling the execution of a CPLEX optimization to write to a script file}
\usage{
makeControlLines(
  parMode = PAR_MODE,
  nThreads = NUM_THREADS,
  timeLimit = TIME_LIMIT,
  probeLevel = PROBE_LEVEL,
  emphasis = EMPHASIS,
  varSelect = VAR_SELECT,
  cutLevel = CUT_LEVEL
)
}
\description{
This function produces instructions controlling the execution of a CPLEX optimization to write to a script file
}
