% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILP.R
\name{makeControlObject}
\alias{makeControlObject}
\title{Produce instructions controlling the execution of a CPLEX optimization}
\usage{
makeControlObject(
  timeLimit = TIME_LIMIT,
  probeLevel = PROBE_LEVEL,
  emphasis = EMPHASIS,
  varSelect = VAR_SELECT,
  cutLevel = CUT_LEVEL
)
}
\description{
Produce instructions controlling the execution of a CPLEX optimization
}
