% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\docType{data}
\name{MY_SEED}
\alias{MY_SEED}
\title{TRUE means optimization mode; FALSE means triage mode}
\format{
An object of class \code{integer} of length 1.
}
\usage{
MY_SEED
}
\description{
TRUE means optimization mode; FALSE means triage mode
}
\keyword{datasets}
