% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\docType{data}
\name{MY_TYPES}
\alias{MY_TYPES}
\title{TRUE means optimization mode; FALSE means triage mode}
\format{
An object of class \code{character} of length 1.
}
\usage{
MY_TYPES
}
\description{
TRUE means optimization mode; FALSE means triage mode
}
\keyword{datasets}
