% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\docType{data}
\name{OPPORTUNISTIC}
\alias{OPPORTUNISTIC}
\title{CPLEX solver settings, high-level}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
OPPORTUNISTIC
}
\description{
CPLEX solver settings, high-level
}
\keyword{datasets}
