% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\docType{data}
\name{OPTIMAL}
\alias{OPTIMAL}
\title{internal names for ILP statuses}
\format{
An object of class \code{character} of length 1.
}
\usage{
OPTIMAL
}
\description{
internal names for ILP statuses
}
\keyword{datasets}
