% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\docType{data}
\name{OPTIMIZATION}
\alias{OPTIMIZATION}
\title{Settings for driver functions in Driver.R and Validation.R}
\format{
An object of class \code{logical} of length 1.
}
\usage{
OPTIMIZATION
}
\description{
Settings for driver functions in Driver.R and Validation.R
}
\keyword{datasets}
