% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{postprocessSolution}
\alias{postprocessSolution}
\title{Postprocess a given solution of an ILP to put it into a convenient format: the phenotype and the clause}
\usage{
postprocessSolution(origPhenotypes, solutionMatrix, type, complement = 0)
}
\description{
Postprocess a given solution of an ILP to put it into a convenient format: the phenotype and the clause
}
