% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\docType{data}
\name{PREFIX1}
\alias{PREFIX1}
\title{prefixes for variable names}
\format{
An object of class \code{character} of length 1.
}
\usage{
PREFIX1
}
\description{
prefixes for variable names
}
\keyword{datasets}
