% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{prepareBoundValues}
\alias{prepareBoundValues}
\title{Prepare the upper cutoffs based on the input, type and objective function}
\usage{
prepareBoundValues(
  extremeValues,
  type,
  objective,
  ns,
  sumYs,
  coeffsTrue,
  coeffsAgree
)
}
\description{
Prepare the upper cutoffs based on the input, type and objective function
}
