% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{prepareCoeffsAndExtraScores}
\alias{prepareCoeffsAndExtraScores}
\title{This function prepares the coefficients and the extra scores based on a specified genotype matrix and objective function.}
\usage{
prepareCoeffsAndExtraScores(
  genotypes,
  objective,
  sumYs,
  ns,
  TPscore = 1,
  TNscore = 1,
  FPscore = -1,
  FNscore = -1
)
}
\description{
This function prepares the coefficients and the extra scores based on a specified genotype matrix and objective function.
}
