% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{prepareExtras}
\alias{prepareExtras}
\title{Prepare the extra information relating to the boundary}
\usage{
prepareExtras(
  extraDefinitions,
  ind,
  goodInds,
  extremeValue,
  ns,
  sumYs,
  baseFilename,
  index = 0L
)
}
\description{
Prepare the extra information relating to the boundary
}
