% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{prepareExtremeValues}
\alias{prepareExtremeValues}
\title{NB: the extreme value is in log-space}
\usage{
prepareExtremeValues(
  extremeValue,
  type,
  singleValues,
  singleBestRatio = Inf,
  useTighterBound = FALSE
)
}
\description{
NB: the extreme value is in log-space
}
