% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{prepareOutput}
\alias{prepareOutput}
\title{This function prepares the final output from a sequence of ILPs, one per SNP, in the form of 2 tables: summary and phenotype}
\usage{
prepareOutput(
  output,
  complexPhenotypes,
  SNP,
  ID,
  type,
  numSegments,
  objective,
  singleRes,
  boundValues,
  ns,
  MH,
  outputAssociations = FALSE
)
}
\description{
This function prepares the final output from a sequence of ILPs, one per SNP, in the form of 2 tables: summary and phenotype
}
