% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{pruneSupersets}
\alias{pruneSupersets}
\title{Return the rows in inputMatrix which are not proper supersets of other rows in the same matrix
Note: assumes, without checking, that inputMatrix does not have any comparable rows (in particular, any duplicates)!}
\usage{
pruneSupersets(inputMatrix)
}
\description{
Return the rows in inputMatrix which are not proper supersets of other rows in the same matrix
Note: assumes, without checking, that inputMatrix does not have any comparable rows (in particular, any duplicates)!
}
