% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{reducePhenotypesAndObjectives}
\alias{reducePhenotypesAndObjectives}
\title{This function reduces an input consisting of a phenotype matrix and a matrix of objective function vectors (same # of rows)
Optionally, it can take a matrix of genotypes and if that input is not NULL, also return a matrix of multiplicities extraDef
The return value is a list containing:
\itemize{
\item the reduced phenotype and objective matrices
\item the status of each row and column: FALSE if it has been removed, TRUE if it has been preserved
\item the vector of extra scores to be added to the objective function
\item optionally, a matrix of extra definitions specifying the multiplicity and total weight of each surviving position
}}
\usage{
reducePhenotypesAndObjectives(phenotypes, objVectors, genotypes = NULL)
}
\description{
This function reduces an input consisting of a phenotype matrix and a matrix of objective function vectors (same # of rows)
Optionally, it can take a matrix of genotypes and if that input is not NULL, also return a matrix of multiplicities extraDef
The return value is a list containing:
\itemize{
\item the reduced phenotype and objective matrices
\item the status of each row and column: FALSE if it has been removed, TRUE if it has been preserved
\item the vector of extra scores to be added to the objective function
\item optionally, a matrix of extra definitions specifying the multiplicity and total weight of each surviving position
}
}
