% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaiIri.R
\name{removeCollinear}
\alias{removeCollinear}
\title{Remove all but the first and the last point of any group of three or more consecutive input points that are collinear.
Note that this function is idempotent (at least in exact arithmetic), i.e. running it twice is the same as running it once.}
\usage{
removeCollinear(pointSet, tol = TOL)
}
\description{
Remove all but the first and the last point of any group of three or more consecutive input points that are collinear.
Note that this function is idempotent (at least in exact arithmetic), i.e. running it twice is the same as running it once.
}
