% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgwas-package.R
\docType{package}
\name{rgwas-package}
\alias{rgwas}
\alias{rgwas-package}
\title{rgwas: Identify Logical Combinations Of Phenotypes Associated With A Genotype In A Multi-Phenotype Genome-Wide Association Study}
\description{
The goal of ReverseGWAS is to find associations between complex phenotypes and genotypes in a multi-phenotype GWAS. If we remove the word 'complex', this becomes identical to the standard GWAS problem of finding genotype-phenotype associations. If we move the word 'complex' next to 'genotypes', this becomes identical to the problem solved by a genetic risk score (GRS). In other words, analogously to a GRS, ReverseGWAS seeks combinations of phenotypes that are simultaneously associated with a SNP. Other than reversing genotype and phenotype, the key difference with a GRS is that ReverseGWAS combines the phenotypes logically.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bichkd/rgwas}
  \item Report bugs at \url{https://github.com/bichkd/rgwas/issues}
}

}
\author{
\strong{Maintainer}: Leonid Chindelevitch \email{leonid.chindelevitch@pfizer.com} (\href{https://orcid.org/0000-0002-6619-6013}{ORCID})

Authors:
\itemize{
  \item Dmitri Bichko \email{dmitri.bichko@pfizer.com}
}

}
\keyword{internal}
