% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\docType{data}
\name{TIME_LIMIT}
\alias{TIME_LIMIT}
\title{CPLEX solver settings, low-level}
\format{
An object of class \code{integer} of length 1.
}
\usage{
TIME_LIMIT
}
\description{
CPLEX solver settings, low-level
}
\keyword{datasets}
