% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\docType{data}
\name{WIDTH}
\alias{WIDTH}
\title{Settings for geometric functions in Geometry.R
Width of the field determining how much we can be off by from the boundary}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
WIDTH
}
\description{
Settings for geometric functions in Geometry.R
Width of the field determining how much we can be off by from the boundary
}
\keyword{datasets}
