% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{applyFormula}
\alias{applyFormula}
\title{Apply a formula in a standard matrix representation (of type "CNF" or "DNF") to a phenotype matrix (assumed to be logical)}
\usage{
applyFormula(parsedFormula, phenotypes, type)
}
\description{
Apply a formula in a standard matrix representation (of type "CNF" or "DNF") to a phenotype matrix (assumed to be logical)
}
