% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logical.R
\name{computeOptValue}
\alias{computeOptValue}
\title{This function computes the optimum value an output shoud have in order to compare it to the actual optimum value}
\usage{
computeOptValue(
  myTab,
  objective,
  coeffsTrue,
  coeffsAgree,
  ind,
  TPscore = 1,
  TNscore = 1,
  FPscore = -1,
  FNscore = -1
)
}
\description{
This function computes the optimum value an output shoud have in order to compare it to the actual optimum value
}
