% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{computeStats}
\alias{computeStats}
\title{Compute a number of 2 x 2 statistics based on a tibble containing at a minimum the columns TP, FP, FN and TN
If pvals = TRUE, additionally compute the Fisher exact and chi-squared p-values (in log-space).}
\usage{
computeStats(Tibble, pvals = TRUE)
}
\description{
Compute a number of 2 x 2 statistics based on a tibble containing at a minimum the columns TP, FP, FN and TN
If pvals = TRUE, additionally compute the Fisher exact and chi-squared p-values (in log-space).
}
