% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logical.R
\name{computeStratifiedPVal}
\alias{computeStratifiedPVal}
\title{Compute the p-value of the Cochran-Mantel-Haenszel test, estimating the association of genotype to phenotype
stratifying by a secondary phenotype; exact = TRUE is a Fisher exact test analog; correct = TRUE is a continuity correction}
\usage{
computeStratifiedPVal(
  genotype,
  phenotypeC,
  phenotypeS,
  exact = FALSE,
  correct = FALSE
)
}
\description{
Compute the p-value of the Cochran-Mantel-Haenszel test, estimating the association of genotype to phenotype
stratifying by a secondary phenotype; exact = TRUE is a Fisher exact test analog; correct = TRUE is a continuity correction
}
