% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{deleteZeroColumns}
\alias{deleteZeroColumns}
\title{This function deletes any all-zero columns of a logical input matrix}
\usage{
deleteZeroColumns(inputMatrix)
}
\description{
This function deletes any all-zero columns of a logical input matrix
}
