% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{extractMinimalColumns}
\alias{extractMinimalColumns}
\title{Extract the minimal columns of a logical input matrix with respect to set inclusion}
\usage{
extractMinimalColumns(inputMatrix)
}
\description{
Extract the minimal columns of a logical input matrix with respect to set inclusion
}
