% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{extractSolution}
\alias{extractSolution}
\title{Extract solution from the output of CPLEX via Rcplex}
\usage{
extractSolution(
  solution,
  prefix1 = PREFIX1,
  prefix2 = PREFIX2,
  optFactor = -1L,
  ignoreNegatives = !is.null(prefix2)
)
}
\description{
Extract solution from the output of CPLEX via Rcplex
}
