% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\name{mainDriver}
\alias{mainDriver}
\title{Run \code{reverseGWAS} on a dataset.}
\usage{
mainDriver(
  inputFile,
  type = "CNF",
  objective = "agreement",
  K = MY_K,
  L = MY_L,
  complement = FALSE,
  ext = stringr::str_sub(inputFile, -4),
  index = 1L,
  extremeValue = log(MAX_P),
  outputSummary = TRUE,
  outputPhenotype = FALSE,
  outputAssociations = FALSE,
  saveOutput = FALSE
)
}
\arguments{
\item{inputFile}{The absolute path to an input file. This can be a CSV or TSV file, either uncompressed or compressed with gz. It must start with a patient identifier column named ID (typically a character), followed by one or more genotype columns, followed by one or more phenotype columns. NB: the file extension must be immediately preceded by an integer stating the number of genotype columns.}

\item{type}{The type of phenotype combinations to search over; the allowed values are "CNF", or conjunctive normal form (the default) and "DNF", or disjunctive normal form.}

\item{objective}{The association function between genotype and complex phenotype to optimise; the allowed values are "agreement" (the default) and "covariance".}

\item{K, L}{The maximum number of clauses (K) and the maximum size of each clause (L) to use in searching for the combination phenotype. The defaults are \code{2} and \code{2}. NB: high values may lead to overfitting to the training (e.g. discovery) data!}

\item{complement}{Integer between \code{-2} and \code{1}, or a value coercible to one. If it is  \code{1} or \code{2}, each genotype's negated form is added in preprocessing. If it is \code{-1} or \code{2}, each phenotype's negated form is added in preprocessing.}

\item{ext}{The extension of the uncompressed input file (CSV or TSV); by default, it is inferred automatically from the input filename.}

\item{index}{Offset value to be used in distinguishing between the iterations; the default is \code{1}.}

\item{extremeValue}{The target value of the association function (by default, the Fisher exact test p-value). The default is log(5e-8), i.e. genome-wide significance. NB: it must be in logarithmic space (i.e. non-positive).}

\item{outputSummary}{Create a file with the summary statistics for the combination phenotype? The default is \code{TRUE}.}

\item{outputPhenotype}{Create a file with the best combination phenotype's value for each patient? The default is \code{FALSE}.}

\item{outputAssociations}{Create a file with the single best phenotype associated to each of the genotypes? The default is \code{FALSE}.}

\item{saveOutput}{Save the output as an RDS file? The default is \code{FALSE}.}
}
\value{
A list containing a summary of the association statistics (one row per genotype) and the best combination phenotypes (one column per genotype).
If outputSummary, outputPhenotype, or outputAssociations is \code{TRUE}, the function also creates an output file containing the summary, the phenotypes, or the single-best associations for each phenotype, respectively.
If saveOutput is \code{TRUE}, the output is also saved as an RDS file.
}
\description{
Main driver for a multi-genotype multi-phenotype association with a single type, number of clauses and clause size.
}
\examples{
# X <- mainDriver(inputFile = "TestInputN5000P5_3.csv", extremeValue = log(1e-3))
}
