% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImaiIri.R
\name{normalizeVector}
\alias{normalizeVector}
\title{This function turns an input 2-dimensional vector into a unit vector in the same direction}
\usage{
normalizeVector(vector)
}
\description{
This function turns an input 2-dimensional vector into a unit vector in the same direction
}
