% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{parseFormula}
\alias{parseFormula}
\title{Parse a formula of type "CNF" or "DNF" into a standard matrix representation}
\usage{
parseFormula(Formula, type)
}
\description{
Parse a formula of type "CNF" or "DNF" into a standard matrix representation
}
