% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{validationDriver}
\alias{validationDriver}
\title{Run \code{reverseGWAS} on a discovery and validation split of the data.}
\usage{
validationDriver(
  inputFile,
  type = "CNF",
  objective = "agreement",
  complement = FALSE,
  extremeValue = log(MAX_P),
  Klist = KLIST,
  Llist = LLIST,
  index = 0,
  outputAssociations = FALSE,
  shuffle = TRUE,
  SEED = 987654321L
)
}
\arguments{
\item{inputFile}{The absolute path to an input file. This can be a CSV or TSV file, either uncompressed or compressed with gz. It must start with a patient identifier column named ID (typically a character), followed by one or more genotype columns, followed by one or more phenotype columns. NB: the file extension must be immediately preceded by an integer stating the number of genotype columns.}

\item{type}{The type of phenotype combinations to search over; the allowed values are "CNF", or conjunctive normal form (the default) and "DNF", or disjunctive normal form.}

\item{objective}{The association function between genotype and complex phenotype to optimise; the allowed values are "agreement" (the default) and "covariance".}

\item{complement}{Integer between \code{-2} and \code{1}, or a value coercible to one. If it is  \code{1} or \code{2}, each genotype's negated form is added in preprocessing. If it is \code{-1} or \code{2}, each phenotype's negated form is added in preprocessing.}

\item{extremeValue}{The target value of the association function (by default, the Fisher exact test p-value). The default is log(5e-8), i.e. genome-wide significance. NB: it must be in logarithmic space (i.e. non-positive).}

\item{Klist, Llist}{Integer vectors of values for \code{K} (number of clauses) and \code{L} (size of clauses); they must have the same length!}

\item{index}{Offset value to be used in distinguishing between the iterations; the default is \code{1}.}

\item{outputAssociations}{Create a file with the single best phenotype associated to each of the genotypes? The default is \code{FALSE}.}

\item{shuffle}{Perform an RCT? If \code{TRUE} (the default), simulate an RCT; otherwise, the first half of the patients (input rows) go into discovery and the second half, into validation.}

\item{SEED}{Integer seed for the random number generator; keeping it consistent between runs generates the same RCT patient split, thus ensuring reproducible results; default: \code{987654321}.}
}
\description{
Carry out the discovery-validation pipeline, whose arguments parallel those of the mainDriver and similar functions.
}
\examples{
# Y <- rgwas::validationDriver(inputFile = "TestInputN500000P10_3.csv.gz", extremeValue = log(5e-8), shuffle = TRUE)

}
